+++ src/fronts/film-today/variantC/components/DefaultCollection.tsx
