+++ src/fronts/media-briefing/variantB/components/CommentCollection.tsx
