+++ src/fronts/media-briefing/variantB/components/DefaultCollection.tsx
