+++ src/fronts/media-briefing/variantB/components/LinkCollection.tsx
