+++ src/fronts/media-briefing/variantB/components/TopCollection.tsx
