+++ src/fronts/media-briefing/variantC/components/CommentCollection.tsx
