+++ src/fronts/media-briefing/variantC/components/DefaultCollection.tsx
