+++ src/fronts/media-briefing/variantC/components/LinkCollection.tsx
