+++ src/fronts/media-briefing/variantC/components/TopCollection.tsx
