+++ src/fronts/opinion/variantB/components/Card.tsx
