+++ src/fronts/opinion/variantB/components/Collection.tsx
