+++ src/fronts/opinion/variantB/components/Grid.tsx
