+++ src/fronts/opinion/variantB/components/LinkCard.tsx
