+++ src/fronts/opinion/variantB/components/MediaCard.tsx
