+++ src/fronts/opinion/variantC/components/Card.tsx
