+++ src/fronts/opinion/variantC/components/Collection.tsx
