+++ src/fronts/opinion/variantC/components/Grid.tsx
