+++ src/fronts/opinion/variantC/components/LinkCard.tsx
