+++ src/fronts/opinion/variantC/components/MediaCard.tsx
